% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_scidb.R
\name{as.tbl_scidb}
\alias{as.tbl_scidb}
\title{Coerce an existing data structure into a `tbl_scidb`}
\usage{
as.tbl_scidb(x, db, ...)
}
\arguments{
\item{x}{an object to convert. Built in methods will convert vectors, arrays,
some sparse matrices and data frames (see \code{\link{as.scidb}}).}

\item{db}{a scidb database connetion object.}

\item{...}{additional parameters passed to \code{\link{as.scidb}}.}
}
\description{
Coerce an existing data structure into a `tbl_scidb`
}
