% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_scidb.R
\name{as_data.frame.tbl_scidb}
\alias{as_data.frame.tbl_scidb}
\title{Download a remote SciDB array into an R data frame}
\usage{
as_data.frame.tbl_scidb(x, ...)
}
\arguments{
\item{x}{an object of class \code{tbl_scidb}}

\item{...}{additional arguments passed to \code{\link{as.R}}}
}
\description{
Download a remote SciDB array into an R data frame
}

