% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_scidb.R
\name{print.tbl_scidb}
\alias{print.tbl_scidb}
\title{Print a summary of a SciDB tbl object}
\usage{
print.tbl_scidb(x, ...)
}
\arguments{
\item{x}{an object of class \code{tbl_scidb}}

\item{...}{additional arguments passed to the print method}
}
\description{
Print a summary of a SciDB tbl object
}

