% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_scidb.R
\name{tbl.scidb}
\alias{tbl.scidb}
\title{Create a dplyr tbl from a \code{scidb} object}
\usage{
\method{tbl}{scidb}(data, ...)
}
\arguments{
\item{data}{an object of class \code{scidb}}
}
\value{
an object of class \code{tbl_scidb}
}
\description{
Create a dplyr tbl from a \code{scidb} object
}
\seealso{
\code{\link{scidb}} \code{\link{tbl_scidb}}
}

