% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_scidb.R
\name{tbl_scidb}
\alias{tbl_scidb}
\title{Create a dplyr tbl from a \code{scidb} object}
\usage{
tbl_scidb(x)
}
\arguments{
\item{x}{an object of class \code{scidb}}
}
\value{
an object of class \code{tbl_scidb}
}
\description{
Create a dplyr tbl from a \code{scidb} object
}
\seealso{
\code{\link{scidb}}
}

